% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneck_methods.R
\name{Beta.coef}
\alias{Beta.coef}
\title{Title}
\usage{
Beta.coef(coef, sig.fit)
}
\arguments{
\item{coef}{double}

\item{sig.fit}{numeric}
}
\value{
matrix
}
\description{
Title
}
\examples{
print('')
}
