% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneck_methods.R
\name{blockGLasso_s}
\alias{blockGLasso_s}
\title{Title}
\usage{
blockGLasso_s(
  X,
  iterations = 2000,
  burnIn = 1000,
  lambdaPriora = 1,
  lambdaPriorb = 1/10,
  verbose = TRUE
)
}
\arguments{
\item{X}{data.frame}

\item{iterations}{interger}

\item{burnIn}{interger}

\item{lambdaPriora}{numeric}

\item{lambdaPriorb}{numeric}

\item{verbose}{logic}
}
\value{
double
}
\description{
Title
}
\examples{
print('ok')
}
