% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectMat.R
\name{calculate_diffBN}
\alias{calculate_diffBN}
\title{Title calculate_diffBN}
\usage{
calculate_diffBN(diffBN, diffBN_mode = "mean")
}
\arguments{
\item{diffBN}{list of diffBN raw result}

\item{diffBN_mode}{character, determine the diffBN function, usually "mean"}
}
\value{
diffBN results
}
\description{
When function "run_diffBN" is done, the ouput of "run_diffBN" can be the input of this function to produce diffBN results of other modes
}
\examples{
calculate_diffBN(test_result,diffBN_mode='mean') # test_result is the result of function "run_diffBN"
}
