% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneck_methods.R
\name{cmi}
\alias{cmi}
\title{Title}
\usage{
cmi(v1, v2, nargin = 2, vcs = NULL)
}
\arguments{
\item{v1}{character}

\item{v2}{character}

\item{nargin}{numeric}

\item{vcs}{numeric}
}
\value{
numeric
}
\description{
Title
}
\examples{
print('')
}
