% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneck_methods.R
\name{cmi2ni}
\alias{cmi2ni}
\title{Title}
\usage{
cmi2ni(dat, lambda, order0 = 0, nargin = 2)
}
\arguments{
\item{dat}{data.frame}

\item{lambda}{double}

\item{order0}{interger}

\item{nargin}{interger}
}
\value{
list
}
\description{
Title
}
\examples{
print('')
}
