% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning.R
\name{combineDAGs}
\alias{combineDAGs}
\title{Title combineDAGs}
\usage{
combineDAGs(dags, Emin = NULL, Emax = NULL)
}
\arguments{
\item{dags}{list of net structures, the result of DG_grid}

\item{Emin}{integer, the minimal arc number. Any net that have arc number lower that Emin will be filtered}

\item{Emax}{integer, the maximal arc number. Any net that have arc number higher that Emax will be filtered}
}
\value{
information of the summarized net structure
}
\description{
combineDAGs is a function that combine all the net structures in the dags according to a filter of Emin and Emax to get the comprehensize network.
}
\examples{
library(foreach)
library(bnlearn)
test <- ma_test[names(ma_test_glist)[1:3000],]
net_struc_ori <- DG_grid(test,params=seq(0.08,0.23,0.01),root='HSC')
net_struc <- combineDAGs(net_struc_ori,Emin=20,Emax=30)
}
