% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{combineDAGsmpl}
\alias{combineDAGsmpl}
\title{combineDAGsmpl}
\usage{
combineDAGsmpl(dags.smpl, Emin = NULL, Emax = NULL, ncores = 64)
}
\arguments{
\item{dags.smpl}{list of net structures, the result of DG_smpl}

\item{Emin}{integer, the minimal arc number. Any net that have arc number lower that Emin will be filtered}

\item{Emax}{integer, the maximal arc number. Any net that have arc number higher that Emax will be filtered}

\item{ncores}{integer, number of cores used to speed up the calculation}
}
\value{
information of the summarized net structure
}
\description{
combineDAGsmpl is a function that combine all the net structures in the dags.smpl according to a filter of Emin and Emax to get the comprehensize network.
}
