% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func.R
\name{dagPlot}
\alias{dagPlot}
\title{Title}
\usage{
dagPlot(DAG, weight = FALSE, ...)
}
\arguments{
\item{DAG}{list}

\item{weight}{logic}

\item{...}{other}
}
\value{
plot
}
\description{
Title
}
\examples{
print('ok')
}
