% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_fibro}
\alias{dat_fibro}
\title{ma_test}
\format{
An object of class \code{data.frame} with 2000 rows and 392 columns.
}
\usage{
dat_fibro
}
\description{
Single cell data of fibroblast differentiation, have 2000 genes and 392 cells
}
\examples{
  head(dat_fibro)
}
\keyword{datasets}
