% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_fibro_meta}
\alias{dat_fibro_meta}
\title{ma_test}
\format{
An object of class \code{character} of length 392.
}
\usage{
dat_fibro_meta
}
\description{
Meta data for single cell data dat_fibro, have 392 cells and 5 cell types in total
}
\examples{
  head(dat_fibro_meta)
}
\keyword{datasets}
