% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneck_methods.R
\name{edgereduce}
\alias{edgereduce}
\title{Title}
\usage{
edgereduce(G, Gval, order, dat, t, lambda, nargin = 2)
}
\arguments{
\item{G}{matrix}

\item{Gval}{matrix}

\item{order}{numeric}

\item{dat}{matrix}

\item{t}{numeric}

\item{lambda}{numeric}

\item{nargin}{numeric}
}
\value{
list
}
\description{
Title
}
\examples{
print('')
}
