% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fibro_glist}
\alias{fibro_glist}
\title{ma_test}
\format{
An object of class \code{character} of length 2000.
}
\usage{
fibro_glist
}
\description{
High-variance genelist for dat_fibro, have 2000 genes in total. Can be calculated by Seurat::FindVariableFeatures
}
\examples{
  head(fibro_glist)
}
\keyword{datasets}
