% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning.R
\name{filterParams}
\alias{filterParams}
\title{Title}
\usage{
filterParams(nEdges, Emin = min(nEdges), Emax = max(nEdges))

filterParams(nEdges, Emin = min(nEdges), Emax = max(nEdges))
}
\arguments{
\item{nEdges}{interger}

\item{Emin}{interger}

\item{Emax}{interger}
}
\value{
interger

interger
}
\description{
Title

Title
}
\examples{
print('ok')
print('ok')
}
