% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectMat.R
\name{get_diffBN}
\alias{get_diffBN}
\title{Title get_diffBN}
\usage{
get_diffBN(scm.inter = NULL, scm.ref = NULL, mode = "mean")
}
\arguments{
\item{scm.inter}{data.frame of bn.fit results}

\item{scm.ref}{data.frame of bn.fit reference, which don't have any gene deletion or permutation}

\item{mode}{character, determine how the diffBN is calculate, usually use "mean"}
}
\value{
data.frame of diffBN
}
\description{
get_diffBN: based on a set of bn.fit results and a bn.fir reference, calculate the diffBN results
}
\examples{
library(foreach)
library(parallel)
library(doParallel)
registerDoParallel(8) # use 8 cores to speed up the calculation
scm.ref = get_scm(mem = ma_test, graph = e, id = "ref") # e is the result of function "trimDAG", is the final netword structure
Features <- rownames(ma_test)
system.time({
  scm.inter = foreach(i = 1:nrow(ma_test), .combine = cbind) \%dopar\% {
    get_scm(ma_test = ma_test[-i,], graph = graph, id = Features[i])
  }
})
diffBN <- get_diffBN(scm.inter = scm.inter, scm.ref = scm.ref)
}
