% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneck_methods.R
\name{glasso_sf}
\alias{glasso_sf}
\title{Title}
\usage{
glasso_sf(X, alpha, niter = 5)
}
\arguments{
\item{X}{data.frame}

\item{alpha}{double}

\item{niter}{interger}
}
\value{
data.frame
}
\description{
Title
}
\examples{
print('')
}
