% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HHATAC_TF}
\alias{HHATAC_TF}
\title{HHATAC_TF}
\format{
An object of class \code{Seurat} with 1764 rows and 1788 columns.
}
\usage{
HHATAC_TF
}
\description{
Single cell atac-seq TF score data of human HSC differentiation axis, have 9 cell types and 2 lineages
}
\examples{
  head(HHATAC_TF)
}
\keyword{datasets}
