% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HHATAC_TF_glist}
\alias{HHATAC_TF_glist}
\title{HHATAC_TF_glist}
\format{
An object of class \code{numeric} of length 1764.
}
\usage{
HHATAC_TF_glist
}
\description{
High-variance TF list for data, have 1764 features in a decreasing order.
}
\examples{
  head(HHATAC_TF_glist)
}
\keyword{datasets}
