% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HMMA}
\alias{HMMA}
\title{HMMA}
\format{
An object of class \code{data.frame} with 21678 rows and 11 columns.
}
\usage{
HMMA
}
\description{
The test data of microarray expression matrix of HSC differentiation, have 11 cell types and 2 lineages
}
\examples{
head(HMMA)
}
\keyword{datasets}
