% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HMMA_glist}
\alias{HMMA_glist}
\title{HMMA}
\format{
An object of class \code{numeric} of length 21679.
}
\usage{
HMMA_glist
}
\description{
High-MI genelist for HMMA. MI can be calculated be infotheo::mutinformation
}
\examples{
  head(HMMA_glist)
}
\keyword{datasets}
