% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneck_methods.R
\name{InvSig.diag.new}
\alias{InvSig.diag.new}
\title{Title}
\usage{
InvSig.diag.new(Y, Beta)
}
\arguments{
\item{Y}{matrix}

\item{Beta}{matrix}
}
\value{
matrix
}
\description{
Title
}
\examples{
print('')
}
