% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneck_methods.R
\name{jsrm}
\alias{jsrm}
\title{Title}
\usage{
jsrm(Y, sig.use, n, p, lam1, lam2, n_iter = 1000)
}
\arguments{
\item{Y}{matrix}

\item{sig.use}{double}

\item{n}{integer}

\item{p}{double}

\item{lam1}{double}

\item{lam2}{double}

\item{n_iter}{integer}
}
\value{
matrix
}
\description{
Title
}
\examples{
print('')
}
