% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning.R
\name{mat2df}
\alias{mat2df}
\title{Title mat2df}
\usage{
mat2df(DAG.dm)
}
\arguments{
\item{DAG.dm}{A matrix that represent a net structure}
}
\value{
Summarized net structure
}
\description{
mat2df: transform a matrix to a summarized net structure
}
\examples{
library(foreach)
library(bnlearn)
test <- ma_test[names(ma_test_glist)[1:3000],]
net_struc_ori <- DG_grid(test,params=seq(0.08,0.23,0.01),root='HSC')
net_struc <- combineDAGs(net_struc_ori,Emin=20,Emax=30)
net_struc_mat <- df2mat(net_struc)
net_struc <- mat2df(net_struc_mat)
}
