% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneck_methods.R
\name{network.bayesianglasso}
\alias{network.bayesianglasso}
\title{Title}
\usage{
network.bayesianglasso(expr.data, prob)
}
\arguments{
\item{expr.data}{matrix}

\item{prob}{numeric}
}
\value{
data.frame
}
\description{
Title
}
\examples{
print('ok')
}
