% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneck_methods.R
\name{network.GeneNet}
\alias{network.GeneNet}
\title{Title}
\usage{
network.GeneNet(expr.data, fdr)
}
\arguments{
\item{expr.data}{data.frame}

\item{fdr}{double}
}
\value{
data.frame
}
\description{
Title
}
\examples{
print('')
}
