% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneck_methods.R
\name{network.glasso}
\alias{network.glasso}
\title{Title}
\usage{
network.glasso(expr.data, lambda)
}
\arguments{
\item{expr.data}{data.frame}

\item{lambda}{double}
}
\value{
data.frame
}
\description{
Title
}
\examples{
print('')
}
