% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneck_methods.R
\name{network.glassosf}
\alias{network.glassosf}
\title{Title}
\usage{
network.glassosf(expr.data, alpha)
}
\arguments{
\item{expr.data}{data.frame}

\item{alpha}{double}
}
\value{
data.frame
}
\description{
Title
}
\examples{
print('')
}
