% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneck_methods.R
\name{pcacmi_edgereduce}
\alias{pcacmi_edgereduce}
\title{Title}
\usage{
pcacmi_edgereduce(G, Gval, order, dat, t, lambda, nargin = 2)
}
\arguments{
\item{G}{1}

\item{Gval}{1}

\item{order}{1}

\item{dat}{1}

\item{t}{1}

\item{lambda}{1}

\item{nargin}{1}
}
\value{
1
}
\description{
Title
}
\examples{
print('')
}
