% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectMat_func.R
\name{run_diffBN}
\alias{run_diffBN}
\title{run_diffBN}
\usage{
run_diffBN(
  net_struc,
  data,
  meta,
  index,
  n_sample = 20,
  n_permutation = 20,
  diffBN_mode = "mean",
  verbal = F,
  replace = F
)
}
\arguments{
\item{net_struc}{bn.fit structure}

\item{data}{data.frame of single cell expression data}

\item{meta}{character of metadata corresponding to expression data}

\item{index}{list of sampling results}

\item{n_sample}{integer, times of cell sampling, usually 20-50}

\item{n_permutation}{integer, times of gene permutation, usually 20-50}

\item{diffBN_mode}{character, determine the diffBN function, usually "mean"}
}
\value{
diffBN results
}
\description{
run_diffBN is a function that calculate diffBN results based on a single cell expression data
}
