% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneck_methods.R
\name{space.joint}
\alias{space.joint}
\title{Title}
\usage{
space.joint(Y.m, lam1, lam2 = 0, sig = NULL, weight = NULL, iter = 2)
}
\arguments{
\item{Y.m}{matrix}

\item{lam1}{double}

\item{lam2}{double}

\item{sig}{NULL}

\item{weight}{NULL}

\item{iter}{integer}
}
\value{
list
}
\description{
Title
}
\examples{
print('')
}
