% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning.R
\name{trimDAG}
\alias{trimDAG}
\title{Title trimDAG}
\usage{
trimDAG(dat_tmp, e, min_arc = 2, max_arc = 4, threshold_value = 0.9)
}
\arguments{
\item{dat_tmp}{matrix of gene expression based on different cell types}

\item{e}{A bn.fit net structure}

\item{min_arc}{integer, the minimal number of arcs pointing to a node}

\item{max_arc}{integer, the maximal number of arcs pointing to a node}

\item{threshold_value}{numeric, determine the filter threshold with which arcs are filtered according to bn.fit results}
}
\value{
A new bn.fit net structure after filter
}
\description{
trimDAG is a function that modify netword structure according to bn.fit results
}
\examples{
library(bnlearn)
e <-  empty.graph(colnames(ma_test))
net_struc <- rmCyc(net_struc) # net_struc is the result of combineDAGs
strc1 <- (net_struc \%>\% rmCyc)[,1:2]
strc1[,1]=as.character(strc1[,1])
strc1[,2]=as.character(strc1[,2])
arcs(e) <-  strc1
trimDAG(ma_test,e,2,4,1)
}
