% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning.R
\name{UG_methods}
\alias{UG_methods}
\title{UG_methods}
\usage{
UG_methods(mem, param = 0.2, plot = FALSE, weight = FALSE, ugMethod = "cmi2ni")
}
\arguments{
\item{mem}{matrix of data, gene as rows and cell types as cols}

\item{param}{numeric, the parameter of UG methods, ranging from 0 to 1}

\item{plot}{logic}

\item{weight}{logic}

\item{ugMethod}{character, determine the method used, usually cmi2ni. Other alternatives including ns/GeneNet/glasso/pcacmi/bayesianglasso}
}
\value{
matrix, return the UG method calculation results
}
\description{
UG_methods is the function that use indirect graph algorithm to get the connection between nodes
}
\examples{
test=ma_test[names(ma_test_glist)[1:3000],]
UG_methods(test,param=0.2)
}
