% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{combineDAGs}
\alias{combineDAGs}
\title{combineDAGs}
\usage{
combineDAGs(dags, Emin = NULL, Emax = NULL)
}
\arguments{
\item{dags}{list of net structures, the result of DG_grid}

\item{Emin}{integer, the minimal arc number. Any net that have arc number lower that Emin will be filtered}

\item{Emax}{integer, the maximal arc number. Any net that have arc number higher that Emax will be filtered}
}
\value{
information of the summarized net structure
}
\description{
combineDAGs is a function that combine all the net structures in the dags according to a filter of Emin and Emax to get the comprehensize network.
}
