% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning.R
\name{combineDAGsmpl}
\alias{combineDAGsmpl}
\title{Title combineDAGsmpl}
\usage{
combineDAGsmpl(dags.smpl, Emin = NULL, Emax = NULL, ncores = 64)
}
\arguments{
\item{dags.smpl}{list of net structures, the result of DG_smpl}

\item{Emin}{integer, the minimal arc number. Any net that have arc number lower that Emin will be filtered}

\item{Emax}{integer, the maximal arc number. Any net that have arc number higher that Emax will be filtered}

\item{ncores}{integer, number of cores used to speed up the calculation}
}
\value{
information of the summarized net structure
}
\description{
combineDAGsmpl is a function that combine all the net structures in the dags.smpl according to a filter of Emin and Emax to get the comprehensize network.
}
\examples{
library(foreach)
library(bnlearn)
library(Seurat)
dat <- CreateSeuratObject(dat_fibro)
dat$celltype <- dat_fibro_meta
Idents(dat)=dat$celltype
a <- DG_smpl(dat[fibro_glist,], frac = 0.1, N_smpl = 10, params = seq(0.001,0.01,0.001),whiteList = NULL, blackList = NULL, root = 'MEF',ugMethod = 'cmi2ni', dagMethod = 'hc', ncores = 64)
b <- combineDAGsmpl(a,Emin=1,Emax=8)
}
