% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning.R
\name{df2mat}
\alias{df2mat}
\title{Title df2mat}
\usage{
df2mat(DAG, ctypes = NULL)
}
\arguments{
\item{DAG}{Summarized net structure}

\item{ctypes}{character, the cell typs used}
}
\value{
A matrix that is corresponding to the summarized net structure
}
\description{
df2mat: transform a summarized net structure to matrix
}
\examples{
library(foreach)
library(bnlearn)
test <- ma_test[names(ma_test_glist)[1:3000],]
net_struc_ori <- DG_grid(test,params=seq(0.08,0.23,0.01),root='HSC')
net_struc <- combineDAGs(net_struc_ori,Emin=20,Emax=30)
df2mat(net_struc)
}
