% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{DG_grid}
\alias{DG_grid}
\title{DG_grid}
\usage{
DG_grid(
  mem,
  params = c(1:100)/100,
  whiteList = NULL,
  blackList = NULL,
  root = NULL,
  ugMethod = "cmi2ni",
  dagMethod = "hc",
  ncores = 1
)
}
\arguments{
\item{mem}{matrix of data, gene as rows and cell types as cols}

\item{params}{numeric, the parameters of UG methods, ranging from 0 to 1}

\item{whiteList}{data.frame, the list of arcs that must exist in the network}

\item{blackList}{data.frame, the list of arcs that can't exist in the network}

\item{root}{character, the root of the graph, which can't have any node pointing to.}

\item{ugMethod}{character, determine the UG_method used, usually cmi2ni. Other alternatives including ns/GeneNet/glasso/pcacmi/bayesianglasso}

\item{dagMethod}{character, determine the method that calculate the arc direction, usually hc}

\item{ncores}{integer,the cores that used to speed up the calculation}
}
\value{
A list of net structures of different parameters
}
\description{
DG_grid is the function that considers all the parameters in "params" to get a group of net structures based on ugMethod and dagMethod
}
