% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectMat.R
\name{EffectMatrix}
\alias{EffectMatrix}
\title{EffectMatrix}
\usage{
EffectMatrix(
  net_struc,
  data,
  meta = NULL,
  index = NULL,
  permutation_time = 20,
  bootstrap_time = 20,
  diffBN_mode = "mean",
  mode = "single_cell",
  ncores = 1
)
}
\arguments{
\item{net_struc}{bn.fit structure}

\item{data}{data.frame of the expression data}

\item{meta}{character of metadata corresponding to expression data}

\item{index}{list of sampling results}

\item{permutation_time}{integer, times of gene permutation, usually 20-50}

\item{bootstrap_time}{integer, times of cell sampling, usually 20-50}

\item{diffBN_mode}{character, determine the diffBN function, usually "mean"}
}
\value{
diffBN results
}
\description{
EffectMatrix is a function that calculate diffBN results based on a given expression data
}
\examples{
library(foreach)
# not run: test_result <- EffectMatrix(net_struc,dat_fibro,dat_fibro_meta,index,permutation_time=20,bootstrap_time=20,diffBN_mode='mean',mode='single_cell)
}
