% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectMat.R
\name{g_per}
\alias{g_per}
\title{Title}
\usage{
g_per(data_per, gene_pos)
}
\arguments{
\item{data_per}{data.frame}

\item{gene_pos}{interger}
}
\value{
data.frame
}
\description{
Title
}
\examples{
print('')
}
