% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning.R
\name{genBList}
\alias{genBList}
\title{Title}
\usage{
genBList(ug, nodes, root = NULL, blackList = NULL)
}
\arguments{
\item{ug}{data.frame}

\item{nodes}{character}

\item{root}{character}

\item{blackList}{data.frame}
}
\value{
data.frame
}
\description{
Title
}
\examples{
print('he')
}
