% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectMat_func.R
\name{get_diffBN}
\alias{get_diffBN}
\title{get_diffBN}
\usage{
get_diffBN(scm.inter = NULL, scm.ref = NULL, mode = "mean")
}
\arguments{
\item{scm.inter}{data.frame of bn.fit results}

\item{scm.ref}{data.frame of bn.fit reference, which don't have any gene deletion or permutation}

\item{mode}{character, determine how the diffBN is calculate, usually use "mean"}
}
\value{
data.frame of diffBN
}
\description{
get_diffBN: based on a set of bn.fit results and a bn.fir reference, calculate the diffBN results
}
