% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{mat2df}
\alias{mat2df}
\title{mat2df}
\usage{
mat2df(DAG.dm)
}
\arguments{
\item{DAG.dm}{A matrix that represent a net structure}
}
\value{
Summarized net structure
}
\description{
mat2df: transform a matrix to a summarized net structure
}
