% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneck_methods.R
\name{network.space}
\alias{network.space}
\title{Use space package to predict graph structure}
\usage{
network.space(expr.data, alpha)
}
\arguments{
\item{expr.data}{data.frame}

\item{alpha}{double}
}
\value{
data.frame
}
\description{
Use space package to predict graph structure
}
