% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning.R
\name{rmCyc}
\alias{rmCyc}
\title{Title rmCyc}
\usage{
rmCyc(dS)
}
\arguments{
\item{dS}{Summarized net structure}
}
\value{
Summarized net structure, but the circles are removed
}
\description{
rmCyc: remove the circles in the network according to the hierarchical structure and net matrix
}
\examples{
rmCyc(net_struc) # net_struc is the result of combineDAGs
}
