% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{rmCycL}
\alias{rmCycL}
\title{rmCycL}
\usage{
rmCycL(dS)
}
\arguments{
\item{dS}{Summarized net structure}
}
\value{
Summarized net structure, but the circles are removed
}
\description{
rmCycL: remove the circles in the network according to the hierarchical structure
}
