% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectMat.R
\name{run_diffBN}
\alias{run_diffBN}
\title{Title run_diffBN}
\usage{
run_diffBN(
  net_struc,
  data,
  meta,
  index,
  permutation_time = 20,
  bootstrap_time = 20,
  diffBN_mode = "mean"
)
}
\arguments{
\item{net_struc}{bn.fit structure}

\item{data}{data.frame of single cell expression data}

\item{meta}{character of metadata corresponding to expression data}

\item{index}{list of sampling results}

\item{permutation_time}{integer, times of gene permutation, usually 20-50}

\item{bootstrap_time}{integer, times of cell sampling, usually 20-50}

\item{diffBN_mode}{character, determine the diffBN function, usually "mean"}
}
\value{
diffBN results
}
\description{
run_diffBN is a function that calculate diffBN results based on a single cell expression data
}
\examples{
library(foreach)
library(parallel)
library(doParallel)
registerDoParallel(8) # use 8 cores to speed up the calculation
index <- bootstrap_index(dat_fibro_meta,bootstrap_times=20,ratio=0.4)
test_result <- run_diffBN(net_struc,dat_fibro,dat_fibro_meta,index,permutation_time=20,bootstrap_time=20,diffBN_mode='mean')
}
