% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{trimDAG}
\alias{trimDAG}
\title{trimDAG}
\usage{
trimDAG(
  dat_tmp,
  e,
  min_arc = 2,
  max_arc = 4,
  threshold_value = 0.9,
  plot = TRUE
)
}
\arguments{
\item{dat_tmp}{matrix of gene expression based on different cell types}

\item{e}{A bn.fit net structure}

\item{min_arc}{integer, the minimal number of arcs pointing to a node}

\item{max_arc}{integer, the maximal number of arcs pointing to a node}

\item{threshold_value}{numeric, determine the filter threshold with which arcs are filtered according to bn.fit results}
}
\value{
A new bn.fit net structure after filter
}
\description{
trimDAG is a function that modify netword structure according to bn.fit results
}
